+++ customize.sh
@@ -41,7 +41,8 @@ REPLACE="
 ##########################################################################################
 
 set_permissions() {
-  [ -d "$MODPATH/system/bin" ] && set_perm_recursive "$MODPATH"/system/bin 0 2000 0755 0755; set_perm_recursive "$VALI"/tools/arm64 0 0 0755 0755
+  [ -f "$MODPATH/system/lib64/libnorgothrond.so" ] && set_perm "$MODPATH"/system/lib64/libnorgothrond.so 0 0 644 u:object_r:system_lib_file:s0; set_perm_recursive "$VALI"/tools/arm64 0 0 0755 0755
+
   # Note that all files/folders in magisk module directory have the $MODPATH prefix - keep this prefix on all of your files/folders
   # Some examples: